makes Section as reusable components
